+++ aa_intel_tool/parser/general.py
@@ -4,7 +4,6 @@ General parser functions
 
 # Standard Library
 import re
-from typing import Optional
 
 # Django
 from django.utils.translation import gettext_lazy as _
@@ -18,12 +17,13 @@ from app_utils.logging import LoggerAddTag
 # AA Intel Tool
 from aa_intel_tool import __title__
 from aa_intel_tool.constants import SUPPORTED_INTEL_TYPES
+from aa_intel_tool.exceptions import ParserError
 from aa_intel_tool.parser.module import chatlist, dscan, fleetcomp
 
 logger = LoggerAddTag(my_logger=get_extension_logger(name=__name__), prefix=__title__)
 
 
-def check_intel_type(scan_data: list) -> Optional[str]:
+def check_intel_type(scan_data: list) -> str:
     """
     Check which intel type we have
 
@@ -40,10 +40,10 @@ def check_intel_type(scan_data: list) -> Optional[str]:
         ):
             return intel_type["parser"]
 
-    return None
+    raise ParserError(message=_("No suitable parser found …"))
 
 
-def parse_intel(form_data: str) -> tuple:
+def parse_intel(form_data: str) -> str:
     """
     Parse intel
 
@@ -56,7 +56,10 @@ def parse_intel(form_data: str) -> tuple:
     scan_data = form_data.splitlines()
 
     if len(scan_data) > 0:
-        intel_type = check_intel_type(scan_data=scan_data)
+        try:
+            intel_type = check_intel_type(scan_data=scan_data)
+        except ParserError as exc:
+            raise ParserError(message=exc.message) from exc
 
         available_parser = {
             "dscan": dscan.parse,
@@ -65,16 +68,18 @@ def parse_intel(form_data: str) -> tuple:
         }
 
         if intel_type in available_parser:
-            new_scan, message = available_parser[intel_type](scan_data=scan_data)
+            try:
+                new_scan = available_parser[intel_type](scan_data=scan_data)
+            except ParserError as exc:
+                # Re-raise the Exception
+                raise ParserError(message=exc.message) from exc
 
-            if new_scan is not None:
-                new_scan.raw_data = form_data
-                new_scan.save()
+            # if new_scan is not None:
+            new_scan.raw_data = form_data
+            new_scan.save()
 
-                return new_scan.hash, message
+            return new_scan.hash
 
-            return None, message
+        raise ParserError(message=_("No suitable parser found …"))
 
-        return None, _("No suitable parser found …")
-
-    return None, _("No data to parse …")
+    raise ParserError(message=_("No data to parse …"))