+++ aa_intel_tool/tests/test_parser.py
@@ -6,6 +6,7 @@ Testing the parsers
 from django.test import TestCase
 
 # AA Intel Tool
+from aa_intel_tool.exceptions import ParserError
 from aa_intel_tool.parser.general import check_intel_type, parse_intel
 from aa_intel_tool.tests.utils import (
     load_chatscan_faulty_txt,
@@ -70,8 +71,8 @@ class TestParserGeneral(TestCase):
 
     def test_check_intel_type_none(self):
         """
-        Test should return 'None' as the expected intel type
-        This happens when ninvalid data has been posted
+        Test should throw a ParserError as the expected intel type
+        This happens when invalid data has been posted
 
         :return:
         :rtype:
@@ -80,14 +81,20 @@ class TestParserGeneral(TestCase):
         form_data = load_chatscan_faulty_txt()
         scan_data = str(form_data).splitlines()
 
-        intel_type = check_intel_type(scan_data=scan_data)
-        expected_intel_type = None
+        expected_exception = ParserError
+        expected_message = "A parser error occurred » No suitable parser found …"
 
-        self.assertEqual(first=intel_type, second=expected_intel_type)
+        with self.assertRaises(ParserError):
+            check_intel_type(scan_data=scan_data)
+
+        with self.assertRaisesMessage(
+            expected_exception=expected_exception, expected_message=expected_message
+        ):
+            check_intel_type(scan_data=scan_data)
 
     def test_parse_intel_with_invalid_form_data(self):
         """
-        Test should return 'None' as parsed intel data for invalid form data
+        Test should return a ParserError as parsed intel data for invalid form data
 
         :return:
         :rtype:
@@ -95,16 +102,20 @@ class TestParserGeneral(TestCase):
 
         form_data = load_chatscan_faulty_txt()
 
-        parsed_intel, message = parse_intel(form_data=form_data)
-        expected_intel_data = None
-        expected_message = "No suitable parser found …"
+        expected_exception = ParserError
+        expected_message = "A parser error occurred » No suitable parser found …"
 
-        self.assertEqual(first=parsed_intel, second=expected_intel_data)
-        self.assertEqual(first=message, second=expected_message)
+        with self.assertRaises(ParserError):
+            parse_intel(form_data=form_data)
+
+        with self.assertRaisesMessage(
+            expected_exception=expected_exception, expected_message=expected_message
+        ):
+            parse_intel(form_data=form_data)
 
     def test_parse_intel_empty_form_data(self):
         """
-        Test should return 'None' as parsed intel data for empty form data
+        Test should throw a ParserError as parsed intel data for empty form data
 
         :return:
         :rtype:
@@ -112,9 +123,13 @@ class TestParserGeneral(TestCase):
 
         form_data = ""
 
-        parsed_intel, message = parse_intel(form_data=form_data)
-        expected_intel_data = None
-        expected_message = "No data to parse …"
+        expected_exception = ParserError
+        expected_message = "A parser error occurred » No data to parse …"
 
-        self.assertEqual(first=parsed_intel, second=expected_intel_data)
-        self.assertEqual(first=message, second=expected_message)
+        with self.assertRaises(expected_exception=expected_exception):
+            parse_intel(form_data=form_data)
+
+        with self.assertRaisesMessage(
+            expected_exception=expected_exception, expected_message=expected_message
+        ):
+            parse_intel(form_data=form_data)